Bug-Fix: X11 freezes the laptop occasionally
There was a small bug that X11 totally freezes my entire setup. The issue is
caused by startx-nvidia when it tries to start X11 on vtty2 without loading the
nvidia driver to the kernel. Previously, I thought this case will not occur as
startx-nvidia uses sudo, which in turns wait for user input. However, this is
not the case as sudo can throw password timeout exception.

For now, autostart startx-nvidia has been disabled.